requirements_modified.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
openai==1.3.0
pandas>=2.0.0,<2.1.0
pytest==7.4.3
matplotlib==3.8.2
numpy>=1.24.0 